refactor: move `filter_order_handler.py` to `common` #265
