+++ test/datagateway_api/icat/filters/test_limit_filter.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 import pytest
 
 from datagateway_api.src.common.exceptions import FilterError
-from datagateway_api.src.datagateway_api.filter_order_handler import FilterOrderHandler
+from datagateway_api.src.common.filter_order_handler import FilterOrderHandler
 from datagateway_api.src.datagateway_api.icat.filters import (
     icat_set_limit,
     PythonICATLimitFilter,
