+++ test/datagateway_api/icat/filters/test_order_filter.py
@@ -2,7 +2,7 @@ import pytest
 from typing_extensions import OrderedDict
 
 from datagateway_api.src.common.exceptions import FilterError
-from datagateway_api.src.datagateway_api.filter_order_handler import FilterOrderHandler
+from datagateway_api.src.common.filter_order_handler import FilterOrderHandler
 from datagateway_api.src.datagateway_api.icat.filters import PythonICATOrderFilter
