+++ test/datagateway_api/icat/filters/test_where_filter.py
@@ -1,7 +1,7 @@
 import pytest
 
 from datagateway_api.src.common.exceptions import BadRequestError, FilterError
-from datagateway_api.src.datagateway_api.filter_order_handler import FilterOrderHandler
+from datagateway_api.src.common.filter_order_handler import FilterOrderHandler
 from datagateway_api.src.datagateway_api.icat.filters import PythonICATWhereFilter
